Finalizando CRUD no banco h2 e na memoria

+++ LOGICA/UnidadeManhattan.java
@@ -2,7 +2,7 @@ package MONITORAMENTO.LOGICA;
 
 public class UnidadeManhattan extends Unidade{
 
-	public UnidadeManhattan(int id, float abcissa, float ordenada, boolean video, boolean termometro, boolean co2, boolean ch4) {
+	public UnidadeManhattan(int id, double abcissa, double ordenada, boolean video, boolean termometro, boolean co2, boolean ch4) {
 		super(id, abcissa, ordenada, video, termometro, co2, ch4);
 	}
